requirements.txt
fastapi
uvicorn
GitPython
sentence-transformers
faiss-cpu
langchain
pydantic
langchain-google-genai
langchain-openai
python-dotenv
